requirements.txt
langgraph==0.2.50
tavily-python==0.5.0
langchain==0.3.7
langchain-anthropic==0.3.0
pandoc==2.4
tldextract==5.1.3
fpdf==1.7.2
flask==3.1.0
fastapi==0.115.5
uvicorn==0.32.0
websockets==14.1
python-multipart==0.0.17
python-dotenv==1.0.1
markdown2==2.5.1
pydantic>=2.4.2